че
че

+++ HashFunc.h
@@ -2,7 +2,8 @@
 #define HASH_FUNC_H
 
 #include <stdio.h>
+#include <stdint.h>
 
-int Hash(void* Arr, size_t ArrSize);
+uint64_t Hash(void* Arr, size_t ArrSize);
 
 #endif
